+++ mipidsi/src/lib.rs
@@ -122,6 +122,8 @@ where
     options: ModelOptions,
     // Current MADCTL value copy for runtime updates
     madctl: dcs::SetAddressMode,
+    // State monitor for sleeping TODO: refactor to a Model-connected state machine
+    sleeping: bool,
 }
 
 impl<DI, M, RST> Display<DI, M, RST>
@@ -259,6 +261,13 @@ where
     }
 
     ///
+    /// Returns `true` if display is currently set to sleep.
+    ///
+    pub fn is_sleeping<D: DelayUs<u32>>(&self) -> bool {
+        self.sleeping
+    }
+
+    ///
     /// Puts the display to sleep, reducing power consumption.
     /// Need to call [Self::wake] before issuing other commands
     ///
@@ -266,6 +275,7 @@ where
         self.dcs.write_command(dcs::EnterSleepMode)?;
         // All supported models requires a 120ms delay before issuing other commands
         delay.delay_us(120_000);
+        self.sleeping = true;
         Ok(())
     }
 
@@ -276,6 +286,7 @@ where
         self.dcs.write_command(dcs::ExitSleepMode)?;
         // ST7789 and st7735s have the highest minimal delay of 120ms
         delay.delay_us(120_000);
+        self.sleeping = false;
         Ok(())
     }
 }